backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
asyncpg==0.29.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
python-telegram-bot==20.7
aiohttp==3.10.5
